Refactor to prepare for hass integration.
Moved the main elkm1.py code to elkm1 directory so that a
services.yaml could be added.

Added elkm1/services.yaml. This can be updated in future when add
additional services.

pylint/flake8 run on all code. Most messages fixed. Some not testable
when the code is outside of the hass code base.

Renamed base elkm1 package to elkm1_lib. This was done as name
conflicts arose when trying to lint.

Refactored all device support. Created a base class ElkDeviceBase
in elkm1/__init__.py to support common ElkM1 device methods and
helpers.

Removed use of discovery_info in creating devices. Chatted @ballob
and agree in principle not needed for what we are doing. He still
needs to see to code and if changes required they should be simple
as there is common helpers to handle creation and the change would
be simple and in a single place.

Simplified the message passing between keypad and alarm_control_panel
domains. No longer uses event bus. Does use elkm1_lib callbacks.

Simplified climate code but most of the device class remained intact.
It's in OK shape but I have not spent much time on that code.

Simplified (broke?; I hope not) device_state_attributes.

Simplified and enhanced config parsing. Note that that hidden
attribute will no longer be supported when hass moves to lovelace.
hidden is still supported in this code base and should be simple
to remove in the future. This is a non-breaking change to existing
configs. It does add additional features. For example and "include"
may not be a comma separated list of includes (the existing
include list still works as before, just the syntax of each
entry is enhanced).

A number of minor updates were made to the elkm1_lib to support
this hass code. Most of the changes are non-breaking (changing
initial values on objects, for example using temperature of -40
instead of None). The attribute changed callback is a breaking
change. An additional parameter was added as the first parameter.
Read the elkm1_lib changelog and commit history to see the changes.